Update license from Apache License 2.0 to MIT
This commit updates the project's license from Apache License 2.0 to MIT. The Apache License 2.0 is a permissive license that allows users to use, modify, and distribute the project's code as long as they give attribution to the original author. However, the MIT license is also a permissive license that has fewer restrictions and is more widely used in the open source community. By updating to the MIT license, the project can be more easily integrated with other open source projects that use the same license. This change does not affect the functionality of the code or how it can be used, but it does provide more flexibility for users who want to modify or distribute the project.